small refactor to split logic
